backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
pyjwt
werkzeug
psycopg2-binary
